Goal Code
Here is the code for our goal which uses a distance sensor to detect when the soccer ball enters the goal. Once the soccer ball enters the goal, we flash all three LED Lights for 5 seconds until the ball is taken out of the goal and the game is reset. Furthermore, this code features a reset button that resets the LED light scoring system when pressed.